adicionando alterações sugeridas pelo revisor
