+++ blocks/modal.css
@@ -110,7 +110,6 @@
   font-weight: 400;
   font-size: 18px;
   line-height: 100%;
-  text-align: center;
   cursor: pointer;
 }
