+++ blocks/profile.css
@@ -41,6 +41,9 @@
   padding-right: 18px;
   padding-bottom: 16px;
   margin: 0;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
 }
 
 .profile__info-btn {
@@ -66,6 +69,9 @@
   font-weight: 400;
   font-size: 18px;
   line-height: 100%;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
 }
 
 .profile__add-btn {
